Finish Modelo and Fabricante, I reached the code standard, it's time to continue the rest

+++ src/main/java/com/jciterceros/locadoracarrosjpa/services/ModeloService.java
@@ -13,13 +13,12 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
+import java.util.*;
 
 @Service
 public class ModeloService {
 
+    public static final String ID_DO_MODELO_NAO_ENCONTRADO = "ID do Modelo não encontrado";
     private final ModelMapper mapper;
     private Boolean mapperAllreadyConfigured = false;
     private final ModeloRepository modeloRepository;
@@ -51,59 +50,62 @@ public class ModeloService {
     public ModeloDTO findById(Long id) {
         return modeloRepository.findById(id)
                 .map(modelo -> mapper.map(modelo, ModeloDTO.class))
-                .orElseThrow(() -> new ResourceNotFoundException("ID do Modelo não encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundException(ID_DO_MODELO_NAO_ENCONTRADO));
     }
 
     @Transactional(readOnly = false)
     public ModeloDTO insert(ModeloDTO modelo) {
 
-        // Verifica se o fabricante existe
         Fabricante fabricante = fabricanteRepository.findById(modelo.getId_fabricante())
                 .orElseThrow(() -> new ResourceNotFoundException("ID do Fabricante não encontrado"));
 
-        // Verifica se já existe um modelo com o mesmo nome para esse fabricante
         boolean exists = modeloRepository.existsByNomeAndFabricanteId(modelo.getNome(), modelo.getId_fabricante());
         if (exists) {
             throw new DatabaseException("Modelo já existe para esse fabricante");
         }
         configureMapper();
 
-//        modelo.setNome_fabricante(fabricante.getNome());
-//        return copyEntityToDto(modeloRepository.save(copyDtoToEntity(modelo)));
         Modelo entity = new Modelo();
         entity.setNome(modelo.getNome());
         entity.setFabricante(fabricante);
-//        mapper.map(modeloRepository.save(entity), modelo);
+
         return mapper.map(modeloRepository.save(entity), ModeloDTO.class);
-     }
-
-    // Update modelo
-    @Transactional(readOnly = false)
-    public ModeloDTO update(Long id, ModeloDTO modelo) {
-        Fabricante fabricante = fabricanteRepository.findById(modelo.getId_fabricante())
-                .orElseThrow(() -> new ResourceNotFoundException("ID do Fabricante não encontrado"));
-
-        Modelo entity = getModelo(id);
-
-        entity.setNome(modelo.getNome());
-        entity.setFabricante(fabricante);
-
-        return copyEntityToDto(modeloRepository.save(entity));
     }
 
-    // delete
+    @Transactional(readOnly = false)
+    public ModeloDTO update(Long id, ModeloDTO modeloDTO) {
+        Fabricante fabricante = fabricanteRepository.findById(modeloDTO.getId_fabricante())
+                .orElseThrow(() -> new ResourceNotFoundException("ID do Fabricante não encontrado"));
+
+        Modelo modelo = modeloRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(ID_DO_MODELO_NAO_ENCONTRADO));
+
+        boolean exists = modeloRepository.existsByNomeAndFabricanteId(modeloDTO.getNome(), modeloDTO.getId_fabricante());
+        if (exists) {
+                throw new DatabaseException("Modelo já existe para esse fabricante");
+        }
+
+        configureMapper();
+
+        modelo.setNome(modeloDTO.getNome());
+        modelo.setFabricante(fabricante);
+
+        return mapper.map(modeloRepository.save(modelo), ModeloDTO.class);
+    }
+
     @Transactional(readOnly = false)
     public void delete(Long id) {
-        Modelo entity = getModelo(id);
+        Modelo entity = modeloRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(ID_DO_MODELO_NAO_ENCONTRADO));
+
+        if(!entity.getCarro().isEmpty()){
+            throw new DatabaseException("Não é possível excluir um modelo que possui carros associados");
+        }
         modeloRepository.deleteById(entity.getId());
     }
 
-    private Modelo getModelo(Long id) {
-        return modeloRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException("ID do Modelo não encontrado"));
-    }
     private void configureMapper() {
-        if (mapperAllreadyConfigured) return;
+        if (Boolean.TRUE.equals(mapperAllreadyConfigured)) return;
         mapper.addMappings(new PropertyMap<Modelo, ModeloDTO>() {
             @Override
             protected void configure() {
@@ -114,26 +116,5 @@ public class ModeloService {
             }
         });
         mapperAllreadyConfigured = true;
-    }
-
-    public Modelo copyDtoToEntity(ModeloDTO dto) {
-        Fabricante fabricante = new Fabricante();
-        fabricante.setId(dto.getId_fabricante());
-        fabricante.setNome(dto.getNome_fabricante());
-
-        Modelo entity = new Modelo();
-        entity.setId(dto.getId());
-        entity.setNome(dto.getNome());
-        entity.setFabricante(fabricante);
-        return entity;
-    }
-
-    public ModeloDTO copyEntityToDto(Modelo entity) {
-        ModeloDTO dto = new ModeloDTO();
-        dto.setId(entity.getId());
-        dto.setNome(entity.getNome());
-        dto.setId_fabricante(entity.getFabricante().getId());
-        dto.setNome_fabricante(entity.getFabricante().getNome());
-        return dto;
     }
 }
